Rename jsb-assets-manager
