+++ engine/jsb-assets-manager.js
